Reduce moveit_cpp to minimal core class
